code_refactoring
Whenever we have duplicate code like the loop that finds the entry in hashtable_update and hashtable_lookup, we should think if there is a better way to write this that would avoid the duplication. We should be able to rewrite these procedures to be shorter by defining a new procedure and rewriting both hashtable_update and hashtable_lookup to use that procedure.
 Modify the code for both hashtable_update and hashtable_lookup to have the same behavior they have now, but using fewer lines of code in each procedure.  You should define a new procedure to help with this. Your new version should have approximately the same running time as the original version, but neither hashtable_update or hashtable_lookup should include any for or while loop, and the block of each procedure should be no more than 6 lines long.

 Your procedures should have the same behavior as the originals.