shifting_all_letters_of_a_string_n_number_of_times
Write a procedure, rotate which takes as its input a string of lower case
letters, a-z, and spaces, and an integer n, and returns the string constructed by shifting each of the letters n steps, and leaving the spaces unchanged. Note that 'a' follows 'z'. You can use an additional procedure if you choose to as long as rotate returns the correct string.
 Note that n can be positive, negative or zero.