Memoization
Memoization is a way to make code run faster by saving previously computed results.  Instead of needing to recompute the value of an
expression, a memoized computation first looks for the value in a cache of pre-computed values.

 Define a procedure, cached_execution(cache, proc, proc_input), that takes in
 three inputs: a cache, which is a Dictionary that maps inputs to proc to
 their previously computed values, a procedure, proc, which can be called by just writing proc(proc_input), and proc_input which is the input to proc.
Your procedure should return the value of the proc with input proc_input, but should only evaluate it if it has not been previously called.