rabbits_multiplying_more_realistic_model_using_recursive_function
A (slightly) more realistic model of rabbit multiplication than the Fibonacci model, would assume that rabbits eventually die. For this question, some rabbits die from month 6 onwards.
Define a procedure rabbits that takes as input a number n, and returns a
number that is the value of the nth number in the rabbit sequence.