looking_up_for_the_best_ranked_page
Define a procedure, lucky_search, that takes as input an index, a ranks dictionary (the result of compute_ranks), and a keyword, and returns the one URL most likely to be the best site for that keyword. If the keyword does not appear in the index, lucky_search should return None.