looking_up_ancestors_in_family_tree
Define a procedure, ancestors(genealogy, person), that takes as its first input a Dictionary in the form given above, and as its second input the name of a person. It should return a list giving all the known ancestors of the input person (this should be the empty list if there are none). The order of the list does not matter and duplicates will be ignored.