removing_tags_and_whitespaces_from_a_page
Write a procedure, remove_tags, that takes as input a string and returns
a list of words, in order, with the tags removed. Tags are defined to be strings surrounded by < >. Words are separated by whitespace or tags.  You may assume the input does not include any unclosed tags, that is,   there will be no '<' without a following '>'.