finding_number_in_list_that_repeats_most
Define a procedure, longest_repetition, that takes as input a  list, and returns the element in the list that has the most consecutive repetitions. If there are multiple elements that have the same number of longest repetitions, the result should be the one that appears first. If the input list is empty, it should return None.