repetition_rate_in_a_given_list_and_the_number_repeated_longest
Define a procedure, longest_repetition, that takes as input a list, and returns the element in the list that has the most consecutive repetitions. If there are multiple elements that  have the same number of longest repetitions, the result should be the one that appears first. If the input list is empty, it should return None.